# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Homer Next-Gen API
  version: "4.0.2-draft"
  description: >
    Draft Next-Gen REST API specification for the new homer-core.
servers:
  - url: "{apiPrefix}/api/v4"
    variables:
      apiPrefix:
        default: ""
        description: Optional API prefix (for example, /homer)
...
security:
  - bearerAuth: []
paths:
...
  /exports:
    post:
      tags: [Exports]
      summary: Create export (PCAP/TEXT)
      operationId: exportsCreate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExportRequest"
      responses:
        "201":
          description: Export created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExportResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/ServerError"
...
  /exports/{exportId}/download:
    get:
      tags: [Exports]
      summary: Download export file
      operationId: exportsDownload
      parameters:
        - $ref: "#/components/parameters/ExportId"
      responses:
        "200":
          description: Export file (binary)
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/ServerError"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    authTokenHeader:
      type: apiKey
      in: header
      name: Auth-Token
  parameters:
...
    ExportId:
      name: exportId
      in: path
      required: true
      schema:
        type: string
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    ServerError:
      description: Server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          $ref: "#/components/schemas/Error"
    Error:
      type: object
      required: [type, title, status]
      properties:
        type:
          type: string
          example: https://api.homer/errors/invalid_argument
        title:
          type: string
          example: Invalid argument
        status:
          type: integer
          example: 400
        detail:
          type: string
          example: Field `from` must be <= `to`
        errors:
          type: array
          items:
            $ref: "#/components/schemas/FieldError"
    FieldError:
      type: object
      properties:
        field:
          type: string
        message:
          type: string
    Meta:
      type: object
      properties:
        request_id:
          type: string
        trace_id:
          type: string
        pagination:
          $ref: "#/components/schemas/Pagination"
        message:
          type: string
    Pagination:
      type: object
      properties:
        limit:
          type: integer
        next_cursor:
          type: string
        prev_cursor:
          type: string
        total:
          type: integer
        has_more:
          type: boolean
...
    ExportRequest:
      type: object
      required: [type, query]
      properties:
        type:
          type: string
          enum: [pcap, text]
        query:
          type: object
    ExportResponse:
      type: object
      properties:
        data:
          type: object
          properties:
            export_id:
              type: string
            status:
              type: string
        meta:
          $ref: "#/components/schemas/Meta"
...